# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API Documentation",
        "description": "Swagger/OpenAPI documentation for the Laravel ecommerce backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/wishlist": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get the authenticated user's wishlist",
                "operationId": "listWishlist",
                "responses": {
                    "200": {
                        "description": "Wishlist list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WishlistListResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Add a product to the wishlist",
                "operationId": "addWishlist",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Added to wishlist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WishlistItemResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/wishlist/{wishlist}": {
            "delete": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Remove an item from the wishlist",
                "operationId": "removeWishlist",
                "parameters": [
                    {
                        "name": "wishlist",
                        "in": "path",
                        "description": "Wishlist item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 201
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed from wishlist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/wishlist/check/{productId}": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Check if a product is in the wishlist",
                "operationId": "checkWishlist",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wishlist status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CheckWishlistResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/products/{product}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get reviews for a product",
                "operationId": "listProductReviews",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated product reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewPageResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Create a review for a purchased product",
                "operationId": "createProductReview",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review posted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reviews/{review}": {
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Update an existing review",
                "operationId": "updateReview",
                "parameters": [
                    {
                        "name": "review",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete a review",
                "operationId": "deleteReview",
                "parameters": [
                    {
                        "name": "review",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "MessageResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Logged out successfully"
                }
            },
...
            "WishlistListResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "data": [
                        {
                            "id": 201,
                            "user_id": 1,
                            "product_id": 10,
                            "created_at": "2026-06-26T10:00:00Z",
                            "product": {
                                "id": 10,
                                "name": "Red Chili",
                                "slug": "red-chili",
                                "price": 2.99,
                                "image": "products/chili.jpg"
                            }
                        }
                    ]
                }
            },
            "WishlistItemResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Added to wishlist",
                    "data": {
                        "id": 201,
                        "user_id": 1,
                        "product_id": 10,
                        "created_at": "2026-06-26T10:00:00Z",
                        "product": {
                            "id": 10,
                            "name": "Red Chili",
                            "slug": "red-chili",
                            "price": 2.99,
                            "image": "products/chili.jpg"
                        }
                    }
                }
            },
            "CheckWishlistResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "in_wishlist": true
                }
            },
...
            "ReviewPageResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "data": {
                        "current_page": 1,
                        "data": [
                            {
                                "id": 5,
                                "user_id": 2,
                                "product_id": 10,
                                "rating": 5,
                                "comment": "Fresh, tasty and delivered fast.",
                                "created_at": "2026-06-26T10:00:00Z",
                                "user": {
                                    "id": 2,
                                    "name": "Jane Doe",
                                    "email": "jane@example.com"
                                }
                            }
                        ],
                        "per_page": 10,
                        "last_page": 1,
                        "total": 1
                    }
                }
            },
            "ReviewRequest": {
                "type": "object",
                "example": {
                    "rating": 5,
                    "comment": "Fresh, tasty and delivered fast."
                }
            },
            "ReviewResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Review posted successfully",
                    "data": {
                        "id": 5,
                        "user_id": 2,
                        "product_id": 10,
                        "rating": 5,
                        "comment": "Fresh, tasty and delivered fast.",
                        "created_at": "2026-06-26T10:00:00Z",
                        "user": {
                            "id": 2,
                            "name": "Jane Doe",
                            "email": "jane@example.com"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use the token returned from login or register in the Authorization header as Bearer <token>.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}